refactor: Repository 인터페이스를 Domain 모듈로 이동
Clean Architecture 원칙에 따라 Repository 패턴을 개선:
- Repository 인터페이스를 Infrastructure에서 Domain으로 이동
- Infrastructure는 Domain 인터페이스를 구현하도록 수정
- Application 레이어가 Domain 인터페이스에만 의존하도록 개선
- ConversationHistory 폴더명을 Conversation으로 변경하여 일관성 확보
- 모든 테스트 코드의 참조 경로 업데이트

변경사항:
- Domain/Repositories/ 인터페이스 생성
- Infrastructure 구현체들이 Domain 인터페이스 참조
- Application 서비스들의 의존성을 Domain으로 변경
- DI 컨테이너 등록 수정
- 테스트 프로젝트 참조 경로 업데이트
