+++ ProjectVG.Domain/Entities/Conversation/ChatRole.cs
