+++ ProjectVG.Domain/Entities/Conversation/ConversationHistory.cs
