+++ ProjectVG.Domain/Repositories/ICharacterRepository.cs
@@ -1,6 +1,6 @@
 using ProjectVG.Domain.Entities.Characters;
 
-namespace ProjectVG.Infrastructure.Persistence.Repositories.Characters
+namespace ProjectVG.Domain.Repositories
 {
     public interface ICharacterRepository
     {
@@ -12,4 +12,4 @@ namespace ProjectVG.Infrastructure.Persistence.Repositories.Characters
         Task<IEnumerable<Character>> GetByUserIdAsync(Guid userId);
         Task<IEnumerable<Character>> GetPublicCharactersAsync();
     }
-} 
+}
