+++ ProjectVG.Infrastructure/Persistence/Repositories/Credit/SqlServerCreditTransactionRepository.cs
@@ -1,6 +1,7 @@
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 using ProjectVG.Domain.Entities.Credits;
+using ProjectVG.Domain.Repositories;
 using ProjectVG.Infrastructure.Persistence.EfCore;
 
 namespace ProjectVG.Infrastructure.Persistence.Repositories.Credit
